requirements.txt
# Core Django packages
Django==3.2.12
djangorestframework==3.12.2
django-admin-autocomplete-filter==0.7.1
django-filter==21.1

# Database
psycopg2-binary==2.9.3

# Web scraping
scrapy>=2.8,<2.12
scrapy-djangoitem==1.1.1
scrapy-playwright==0.0.44
playwright

# Environment and utilities
python-dotenv==0.19.2
requests==2.25.1

# Testing (optional)
coverage==6.2